move get_save_and_finalize_callbacks to async_utils
